fix CI errors
