+++ index/src/lib.rs
@@ -5,8 +5,8 @@ extern crate alloc;
 mod constants;
 pub mod executors;
 pub mod prelude;
-pub mod subrpc;
 pub mod registry;
-mod traits;
+pub mod subrpc;
+pub mod traits;
 mod transactors;
 pub mod utils;
