+++ index/src/prelude.rs
@@ -1,4 +1,4 @@
-pub type Evm2PhalaExecutor = super::executors::bridge_executor::Evm2PhalaExecutor;
+pub type Evm2PhalaExecutor = super::executors::bridge_executors::Evm2PhalaExecutor;
 pub use crate::traits::common::Error;
 pub use crate::traits::executor::Executor;
 pub use crate::traits::registry::*;
