+++ index/src/subrpc/ss58.rs
@@ -1,4 +1,4 @@
-use crate::traits::Error;
+use crate::traits::common::Error;
 use alloc::string::String;
 use alloc::vec;
 use base58::ToBase58;
