+++ index/src/traits/common.rs
@@ -1,4 +1,4 @@
-use primitive_types::{H160, H256};
+use primitive_types::{H160, H256, U256};
 use scale::{Decode, Encode};
 
 #[derive(Clone, Encode, Decode, Eq, PartialEq, Debug)]
@@ -15,6 +15,12 @@ pub enum Error {
     ConstructContractFailed,
     FetchDataFailed,
     Unimplemented,
+    InvalidMultilocation,
+    InvalidAmount,
+    SubRPCRequestFailed,
+    InvalidBody,
+    InvalidSignature,
+    Ss58,
 }
 
 #[derive(Clone, Encode, Decode, Eq, PartialEq, Debug)]
@@ -22,4 +28,9 @@ pub enum Error {
 pub enum Address {
     EthAddr(H160),
     SubAddr(H256),
+}
+
+pub enum Amount {
+    U256(U256),
+    U128(u128),
 }
